% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4476.R
\name{add_d4476_plot_details}
\alias{add_d4476_plot_details}
\title{Add D4476 Plot Details}
\usage{
add_d4476_plot_details(results_df, plotlist, i, invert = F)
}
\arguments{
\item{results_df}{the data frame containing the results from D4476 Analysis.}

\item{plotlist}{the list of plots to add to.}

\item{invert}{default is False, when True a plot of deflection vs. load is returned.}
}
\value{
A ggplot of the test with additional points for modulus and failure.
}
\description{
Adds failure point as well as a line between start and stop of modulus slope to a ggplot of load vs. deflection.
}
