% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4476.R
\name{crack.detect}
\alias{crack.detect}
\title{Crack Detection}
\usage{
crack.detect(data, threshold = 2, after = 100)
}
\arguments{
\item{data}{the data frame containing the load, deflection, and time data from test.}

\item{threshold}{default = 2, the drop in load to classify data point as a crack.}

\item{after}{default = 100, the load after which to begin looking for cracks.}
}
\value{
A data frame of places where possible cracks occurred.
}
\description{
Find possible cracking prior to structural failure in results from mechanical testing.
}
