% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4476.R
\name{d4476.analysis}
\alias{d4476.analysis}
\title{D4476 Analysis}
\usage{
d4476.analysis(file_list, measurements, thresh = 2)
}
\arguments{
\item{data.file}{the input file path to the MTS data file.}

\item{meaurements}{the data frame containing the specimen name, Span Length, original diameter, and thickness measurements for each specimen.}
}
\value{
A data frame of results for the test.
}
\description{
Analyze data from an MTS data file to produce data frame of results required by D4476.
}
