% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4476.R
\name{d4476.plots}
\alias{d4476.plots}
\title{D4476 Plots}
\usage{
d4476.plots(results_df, invert = F)
}
\arguments{
\item{data.file}{the input file path to the MTS data file.}

\item{results}{the data frame containing the results from D4476 Analysis.}
}
\value{
A ggplot of the test.
}
\description{
Produce a list of plots for D4476 Testing with highlighted Modulus and Failure points.
}
