% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{get_data_files}
\alias{get_data_files}
\title{Get Data Files}
\usage{
get_data_files(caption = "Select a directory")
}
\arguments{
\item{No}{inputs.}
}
\value{
A dataframe of sample names and dataframes containing raw data for each sample.
}
\description{
Interactively select directory containing data files, returns a dataframe of sample names and dataframes containing raw data.
}
