% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{get_failure}
\alias{get_failure}
\title{Get Failure Points for All Data}
\usage{
get_failure(result_df)
}
\arguments{
\item{result_df:}{The data frame containing data frames of test data.}
}
\value{
Data frame with failure points for each specimen.
}
\description{
Retrieve the maximum load as well as the deflection and time at maximum load for each specimen in a dataframe containing dataframes of specimen data.
}
