% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{get_sample_name}
\alias{get_sample_name}
\title{Get Sample Name}
\usage{
get_sample_name(path)
}
\arguments{
\item{path:}{the path to the MTS data file from which to extract the sample name.}
}
\value{
The sample name.
}
\description{
Extract the sample name from an MTS data file.
}
