% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4476.R
\name{left}
\alias{left}
\title{Left}
\usage{
left(x, n)
}
\arguments{
\item{x}{the string.}

\item{n}{the number of characters to keep.}
}
\value{
String of first n characters of string x.
}
\description{
Return the first n characters from a string.
}
