% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{load_defl_plot}
\alias{load_defl_plot}
\title{Load vs. Deflection Scatter Plot.}
\usage{
load_defl_plot(results_df, invert = F)
}
\arguments{
\item{result_df:}{The data frame containing data frames of test data.}

\item{invert:}{Default is False. When true, a plot with Load on the x-axis and Deflection on y-axis is returned.}
}
\value{
A list of ggplot objects.
}
\description{
Produce a ggplot scatter plot of Load vs. Deflection for all data in a list of load deflection data sets.
}
