% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{sort_plots}
\alias{sort_plots}
\title{Sort Plots}
\usage{
sort_plots(plots)
}
\arguments{
\item{plots:}{a list of ggplots.}
}
\value{
A list of ggplots sorted by plot title.
}
\description{
Sort a list of ggplots by plot title.
}
