% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{span_data}
\alias{span_data}
\title{Span Data}
\usage{
span_data(data, ref, span = -1)
}
\arguments{
\item{data}{the data set}

\item{ref}{the reference point}

\item{span}{the amount you want to span in percent.}
}
\description{
Finds the data point in a data set which is some percentage different from a given reference point.
e.g. if the reference point is 100 lbf, and the span is set to -5, the first point which is greater than or equal to 95\% of the reference (95 lbf) is returned.
}
