% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{right}
\alias{right}
\title{Right}
\usage{
right(x, n)
}
\arguments{
\item{x}{the string.}

\item{n}{the number of characters to keep.}
}
\value{
String of last n characters of string x.
}
\description{
Return the last n characters from a string.
}
